Add tests for multiplication and divide

+++ src/main/java/com/codingblackfemales/Calculator.java
@@ -2,18 +2,22 @@ package com.codingblackfemales;
 
 public class Calculator {
     public static Integer add(final int firstInput, final int secondInput) {
+        
         return firstInput + secondInput;
     }
 
     public static Integer subtract(final int firstInput, final int secondInput) {
+        
         return firstInput - secondInput;
     }
 
     public static Integer multiply(final int firstInput, final int secondInput) {
+        
         return firstInput * secondInput;
     }
 
-    public static Integer divide(final int firstInput, final int secondInput) {
+    public static Float divide(final float firstInput, final float secondInput) {
+        
         return firstInput / secondInput;
     }
 }
